Move wizard-related states into the wizard component
