+++ src/components/wizard/wizard-step.tsx
@@ -3,7 +3,7 @@ import { Link, useLocation } from 'react-router-dom';
 import { Button } from '../ui/button';
 
 interface WizardStepProps {
-    items: {
+    steps: {
         title: string;
         path: string;
         description: string;
@@ -11,17 +11,17 @@ interface WizardStepProps {
     goToStep: (step: number) => void;
 }
 
-const WizardStep = ({ items, goToStep }: WizardStepProps) => {
+const WizardStep = ({ steps, goToStep }: WizardStepProps) => {
     const { pathname } = useLocation();
     return (    
         <nav
             className="grid grid-cols-3"
         >
-            {items.map((item, index) => {
-                const isActive = pathname === item.path;
+            {steps.map((step, index) => {
+                const isActive = pathname === step.path;
                 return (
                     <Button
-                        key={item.title}
+                        key={step.title}
                         variant="ghost"
                         className={`flex gap-4 justify-start hover:invert ${isActive && 'pointer-events-none'}`}
                         onClick={() => goToStep(index)}
@@ -29,9 +29,9 @@ const WizardStep = ({ items, goToStep }: WizardStepProps) => {
                         <div
                             className={`rounded-full w-8 h-8 flex items-center justify-center ${isActive ? `bg-active text-white` : `bg-light-2`}`}
                         >
-                            <p className="base-medium">{items.indexOf(item) + 1}</p> 
+                            <p className="base-medium">{index + 1}</p> 
                         </div>
-                        <p className={`base-medium ${isActive && `text-active base-semibold`}`}>{item.title}</p>
+                        <p className={`base-medium ${isActive && `text-active base-semibold`}`}>{step.title}</p>
                     </Button>
                 )
             })}
